refactor(auto-review): zera react-doctor em AutoReviewPage (Onda 4, #249)
Elimina os 5 diagnosticos do react-doctor em AutoReviewPage.tsx
refatorando (sem supressao), parte da campanha "Zerar react-doctor"
(epico #239), gemeo do piloto MyVerdictsView (#232).

- choiceKey -> escopo de modulo (lib/auto-review-decided.ts, puro e
  ja compartilhado). Mata prefer-module-scope-pure-function.
- Extrai AutoReviewEmptyState, AutoReviewPageHeader e
  AutoReviewPageContent. O pai cai de 341 -> 147 linhas e de 5 -> 3
  useState. Mata no-giant-component + prefer-useReducer (sem
  useReducer, que nao existe no repo).
- Reset de fieldIndex ao trocar de doc via key={doc.docId} no
  AutoReviewPageContent (remontagem), no lugar do useEffect. Mata
  no-adjust-state-on-prop-change (error) + no-derived-state-effect.
  Escolhi key em vez de ajuste-durante-render porque a 0.5.8 marca o
  prev-prop em useState e o eslint barra prev-prop em useRef; key e a
  unica saida limpa nos dois linters. O ResizablePanelGroup fica no
  pai, fora da remontagem, preservando o tamanho do split. Keyar por
  docId (nao indice) ainda reseta quando um doc resolvido sai da fila.
- Memoiza fieldStatus/contagens (qualidade).

Comportamento preservado. AutoReviewFieldPanel.tsx (3 diagnosticos
pre-existentes, incl. o effect :97) fica fora de escopo. API publica
(AutoReviewPage/AutoReviewPageProps) inalterada.

Verificacao: react-doctor 0 diagnosticos nos 4 arquivos (global
errors 7->6, warnings 107->103); tsc 0; eslint 0; vitest 578/578.

Closes #249
